[ANDROSDK-566] Move filter definitions to interfaces
